Fix version schema to be PEP 440 compliant
PEP 440 requires the local part to be delimited by a plus sign. It also
doesn't allow minus signs. So instead of just stripping the 'v' from the
`git describe` we replace all '-' with '_', replace the first '_' with
a '+' und then all others with a dot ('.').

This then might look a little something like '0.5+1.g6ebac70.dirty'.