docs: added Sphinx documentation with manpage
The documentation is not complete yet, but there are Makefile targets to
create it automatically and most sections already exists and just have to
be fleshed out. The cookbook from issue #3 is still completly missing.
The implicit variables from issue #9 are atleast mentioned in the
internal API docstrings.

I reformatted the README.md into reStructuredText so I could easily use
it in the documentation as well. Since GitHub supports rst for READMEs
too, this shouldn't be such a big issue.