Enable Jinja loopcontrols and do extensions
These are built in to Jinja and don't modify existing behaviour.

The [do extension] adds a do statement that can be used to call
arbitrary functions with out the need to assign or print a value. This
can come in handy when pre-processing the data.

The loopcontrols extension adds a continue and break statement that can
be used in for and while loops the same way as in Python or other
programming languages. This doesn't really add to the maintainability of
templates, but can still be helpful for template designers.

[do extension]: http://jinja.pocoo.org/docs/2.10/templates/#expression-statement
[loopcontrols extension]: http://jinja.pocoo.org/docs/2.10/templates/#loop-controls